+++ projects/ngx-scrollbar/src/lib/tests/global-options.spec.ts
@@ -5,9 +5,8 @@ describe('Global options', () => {
   let component: NgScrollbar;
   let fixture: ComponentFixture<NgScrollbar>;
 
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [NgScrollbar],
+  beforeEach(() => {
+    TestBed.configureTestingModule({
       providers: [
         provideScrollbarOptions({
           visibility: 'visible',