+++ projects/ngx-scrollbar/src/lib/viewport/index.ts
@@ -1,2 +1,3 @@
-export * from './scroll-viewport';
 export * from './viewport-adapter';
+export * from './scrollbar-content';
+export * from './scrollbar-viewport';