+++ projects/ngx-scrollbar/src/lib/viewport/scrollbar-content.ts
@@ -4,11 +4,12 @@ import { ChangeDetectionStrategy, Component } from '@angular/core';
   host: {
     '[class.ng-scroll-content]': 'true'
   },
-  selector: 'scroll-content',
+  selector: 'ng-scroll-content',
   template: '<ng-content/>',
   styles: [`
+    /** Make sure content is positioned relatively so scrollbars can be sticky */
     :host {
-      display: block;
+      position: relative;
     }
   `],
   changeDetection: ChangeDetectionStrategy.OnPush