+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-## 18.1.0
+## 19.0.0
 
 - feat: Add `ScrollbarAnywhere` service which provides `createScrollbar` to attach scrollbar on any host element without having to use `ng-scrollbar` component.
 - feat: Add `matSelectScrollbar` directive that can be used to attach scrollbar on `mat-select` overlay menu.
 - refactor: Remove `standalone` from all components.
+
+### Breaking Changes
+
+- `scrollViewport` directive is removed, use the selector on `externalViewport` input instead.
+- When using `externalViewport`, setting the value is now required.
 
 ## 18.0.0